requirements.txt
Flask>=1.0,<=1.1.2
numpy
pandas
neo4j==4.3.4
matplotlib
py2neo
ipython
graphistry

beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
cryptography==36.0.2
html5lib==1.1
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
pycparser==2.21
PyJWT==2.3.0
requests==2.27.1
six==1.16.0
soupsieve==2.3.1
urllib3==1.26.9
webencodings==0.5.1
Werkzeug==2.1.1
xmltodict==0.12.0
zipp==3.8.0